Progress on Ranking data UI
