+++ src/Gateways/Desktop/Core/Components/Styles/TextBlock.xaml
@@ -7,6 +7,10 @@
         <Setter Property="Padding" Value="10px 0 0 0" />
     </Style>
 
+    <Style x:Key="Text-Small" TargetType="TextBlock">
+        <Setter Property="FontSize" Value="12" />
+    </Style>
+
     <Style x:Key="Interval" TargetType="TextBlock">
         <Setter Property="Text" Value="   " />
     </Style>
