+++ src/Gateways/Desktop/Core/Services/IExplorerService.cs
@@ -1,6 +1,6 @@
 ﻿using EnduranceJudge.Core.ConventionalServices;
 
-namespace EnduranceJudge.Gateways.Desktop.Core.Static
+namespace EnduranceJudge.Gateways.Desktop.Core.Services
 {
     public interface IExplorerService : IService
     {
