+++ src/Gateways/Desktop/Views/Content/Common/Participants/ParticipantTemplate.xaml
@@ -1,0 +1,123 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:participations="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Common.Participants"
+                    xmlns:localization="clr-namespace:EnduranceJudge.Localization.Translations;assembly=EnduranceJudge.Localization">
+
+    <DataTemplate DataType="{x:Type participations:ParticipantTemplateModel}">
+        <Border Style="{StaticResource Border-Vertical}">
+            <StackPanel Style="{StaticResource List-Vertical}">
+                <!-- Columns (Headers) -->
+                <StackPanel Style="{StaticResource List-Horizontal}">
+
+                    <!-- Number -->
+                    <Border Style="{StaticResource Border-Horizontal}" Width="150px">
+                        <TextBlock Style="{StaticResource Text-H2}">
+                            <Run>#</Run>
+                            <Run Text="{Binding Number, Mode=OneWay}" />
+                        </TextBlock>
+                    </Border>
+
+                    <!-- Phases -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <ListBox Style="{StaticResource ListBox-Horizontal}" BorderThickness="0"
+                                 ItemsSource="{Binding PhaseLengths}">
+                            <ListBox.ItemTemplate>
+                                <DataTemplate>
+                                    <Border Style="{StaticResource Border-Participation-In-Phase}">
+                                        <TextBlock Style="{StaticResource Text-Centered}" FontWeight="Bold">
+                                            <Run Text="{Binding .}"/>
+                                            <Run Text="{x:Static localization:Words.KM}"/>
+                                        </TextBlock>
+                                    </Border>
+                                </DataTemplate>
+                            </ListBox.ItemTemplate>
+                        </ListBox>
+                    </Border>
+
+                    <!-- Collapse/Expand Button -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <Button Style="{StaticResource Button-Horizontal}"
+                                Command="{Binding ToggleVisibility}"
+                                Content="{Binding ToggleText}" />
+                    </Border>
+                </StackPanel>
+
+                <!-- Performances -->
+                <StackPanel Style="{StaticResource List-Horizontal}" Visibility="{Binding Visibility}">
+
+                    <!-- Labels -->
+                    <Border Style="{StaticResource Border-Horizontal}" Width="150px">
+                        <StackPanel Style="{StaticResource List-Vertical}">
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.START}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.ARRIVE}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.INSPECTION}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.RE_INSPECTION}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.REQUIRED_INSPECTION}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.LOOP_SPAN}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.PHASE_SPAN}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.AVERAGE_SPEED_FOR_LOOP}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.AVERAGE_SPEED_FOR_PHASE}"/>
+                            </Border>
+
+                            <Border Style="{StaticResource Border-Performance-Entry}">
+                                <TextBlock Style="{StaticResource Text-Bold}"
+                                           Text="{x:Static localization:Words.NEXT_START_TIME}"/>
+                            </Border>
+                        </StackPanel>
+                    </Border>
+
+                    <!-- Performances -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <ListBox Style="{StaticResource ListBox-Horizontal}" BorderThickness="0"
+                                 ItemsSource="{Binding Performances}"
+                                Name="Performances" />
+                    </Border>
+
+                    <!-- Print button -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <Button Style="{StaticResource Button-Big-Horizontal}"
+                                Command="{Binding Print}"
+                                CommandParameter="{Binding ElementName=Performances}"
+                                Content="{x:Static localization:Words.PRINT}"/>
+                    </Border>
+                </StackPanel>
+            </StackPanel>
+        </Border>
+    </DataTemplate>
+
+</ResourceDictionary>
