+++ src/Gateways/Desktop/Views/Content/Common/PhasePerformances/PerformanceTemplate.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Manager.PhasePerformances"
+                    xmlns:local="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Common.PhasePerformances"
                     xmlns:localization="clr-namespace:EnduranceJudge.Localization.Translations;assembly=EnduranceJudge.Localization">
 
     <DataTemplate DataType="{x:Type local:PerformanceTemplateModel}">
