+++ src/Gateways/Desktop/Views/Content/Configuration/Children/Competitions/AddParticipants/AddParticipantsView.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using EnduranceJudge.Gateways.Desktop.Core;
-using EnduranceJudge.Gateways.Desktop.Core.Static;
+using EnduranceJudge.Gateways.Desktop.Core.Services;
 using System.Windows.Controls;
 using System.Windows.Input;
 
@@ -7,9 +7,9 @@ namespace EnduranceJudge.Gateways.Desktop.Views.Content.Configuration.Children.C
 
 public partial class AddParticipantsView : UserControl, IScrollableView
 {
-    private readonly IMouseInputService mouseInput;
+    private readonly IMouseHandler mouseInput;
 
-    public AddParticipantsView(IMouseInputService mouseInput) : this()
+    public AddParticipantsView(IMouseHandler mouseInput) : this()
     {
         this.mouseInput = mouseInput;
     }
