+++ src/Gateways/Desktop/Views/Content/Configuration/Children/Competitions/CompetitionView.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using EnduranceJudge.Gateways.Desktop.Core;
-using EnduranceJudge.Gateways.Desktop.Core.Static;
+using EnduranceJudge.Gateways.Desktop.Core.Services;
 using System.Windows.Controls;
 using System.Windows.Input;
 
@@ -7,14 +7,14 @@ namespace EnduranceJudge.Gateways.Desktop.Views.Content.Configuration.Children.C
 {
     public partial class CompetitionView : UserControl, IScrollableView
     {
-        private readonly IMouseInputService mouseInput;
+        private readonly IMouseHandler mouseInput;
 
         public CompetitionView()
         {
             InitializeComponent();
         }
 
-        public CompetitionView(IMouseInputService mouseInput) : this()
+        public CompetitionView(IMouseHandler mouseInput) : this()
         {
             this.mouseInput = mouseInput;
         }
