+++ src/Gateways/Desktop/Views/Content/Configuration/Children/Personnel/PersonnelView.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using EnduranceJudge.Gateways.Desktop.Core;
-using EnduranceJudge.Gateways.Desktop.Core.Static;
+using EnduranceJudge.Gateways.Desktop.Core.Services;
 using System.Net;
 using System.Windows.Controls;
 using System.Windows.Input;
@@ -8,14 +8,14 @@ namespace EnduranceJudge.Gateways.Desktop.Views.Content.Configuration.Children.P
 {
     public partial class PersonnelView : UserControl, IScrollableView
     {
-        private readonly IMouseInputService mouseInput;
+        private readonly IMouseHandler mouseInput;
 
         public PersonnelView()
         {
             InitializeComponent();
         }
 
-        public PersonnelView(IMouseInputService mouseInput) : this()
+        public PersonnelView(IMouseHandler mouseInput) : this()
         {
             this.mouseInput = mouseInput;
         }
