+++ src/Gateways/Desktop/Views/Content/Configuration/Roots/Athletes/Listing/AthleteListView.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using EnduranceJudge.Gateways.Desktop.Core;
-using EnduranceJudge.Gateways.Desktop.Core.Static;
+using EnduranceJudge.Gateways.Desktop.Core.Services;
 using System.Windows.Controls;
 using System.Windows.Input;
 
@@ -7,14 +7,14 @@ namespace EnduranceJudge.Gateways.Desktop.Views.Content.Configuration.Roots.Athl
 {
     public partial class AthleteScrollableView : UserControl, IScrollableView
     {
-        private readonly IMouseInputService mouseInput;
+        private readonly IMouseHandler mouseInput;
 
         public AthleteScrollableView()
         {
             InitializeComponent();
         }
 
-        public AthleteScrollableView(IMouseInputService mouseInput) : this()
+        public AthleteScrollableView(IMouseHandler mouseInput) : this()
         {
             this.mouseInput = mouseInput;
         }
