+++ src/Gateways/Desktop/Views/Content/Configuration/Roots/Participants/ParticipantView.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using EnduranceJudge.Gateways.Desktop.Core;
-using EnduranceJudge.Gateways.Desktop.Core.Static;
+using EnduranceJudge.Gateways.Desktop.Core.Services;
 using System.Windows.Controls;
 using System.Windows.Input;
 
@@ -7,14 +7,14 @@ namespace EnduranceJudge.Gateways.Desktop.Views.Content.Configuration.Roots.Part
 {
     public partial class ParticipantView : UserControl, IScrollableView
     {
-        private readonly IMouseInputService mouseInput;
+        private readonly IMouseHandler mouseInput;
 
         public ParticipantView()
         {
             InitializeComponent();
         }
 
-        public ParticipantView(IMouseInputService mouseInput) : this()
+        public ParticipantView(IMouseHandler mouseInput) : this()
         {
             this.mouseInput = mouseInput;
         }
