+++ src/Gateways/Desktop/Views/Content/Manager/ParticipantsList/ParticipantListView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="EnduranceJudge.Gateways.Desktop.Views.Content.Manager.Participants.Listing.ParticipantListView"
+﻿<UserControl x:Class="EnduranceJudge.Gateways.Desktop.Views.Content.Manager.ParticipantsList.ParticipantListView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:prism="http://prismlibrary.com/"
