+++ src/Gateways/Desktop/Views/Content/Manager/ParticipantsList/ParticipantListView.xaml.cs
@@ -1,15 +1,15 @@
 ﻿using EnduranceJudge.Gateways.Desktop.Core;
-using EnduranceJudge.Gateways.Desktop.Core.Static;
+using EnduranceJudge.Gateways.Desktop.Core.Services;
 using System.Windows.Controls;
 using System.Windows.Input;
 
-namespace EnduranceJudge.Gateways.Desktop.Views.Content.Manager.Participants.Listing
+namespace EnduranceJudge.Gateways.Desktop.Views.Content.Manager.ParticipantsList
 {
     public partial class ParticipantListView : UserControl, IScrollableView
     {
-        private readonly IMouseInputService mouseInput;
+        private readonly IMouseHandler mouseInput;
 
-        public ParticipantListView(IMouseInputService mouseInput) : this()
+        public ParticipantListView(IMouseHandler mouseInput) : this()
         {
             this.mouseInput = mouseInput;
         }
