+++ src/Gateways/Desktop/Views/Content/Rankings/CompetitionListView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="EnduranceJudge.Gateways.Desktop.Views.Content.Rankings.RankingView"
+﻿<UserControl x:Class="EnduranceJudge.Gateways.Desktop.Views.Content.Rankings.CompetitionListView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:prism="http://prismlibrary.com/"
