+++ src/Gateways/Desktop/Views/Content/Rankings/RankListEntry/RankListEntryTemplate.xaml
@@ -1,13 +1,62 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:participations="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Manager.Participants"
-                    xmlns:localization="clr-namespace:EnduranceJudge.Localization.Translations;assembly=EnduranceJudge.Localization">
+                    xmlns:participations="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Rankings.RankListEntry">
 
-    <DataTemplate DataType="{x:Type participations:ParticipantTemplateModel}">
+    <DataTemplate DataType="{x:Type participations:RankListEntryTemplateModel}">
         <Border Style="{StaticResource Border-Vertical}">
             <StackPanel Style="{StaticResource List-Vertical}">
+                <!-- Participant data -->
+                <StackPanel Style="{StaticResource List-Horizontal}">
+                    <!-- Rank -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <TextBlock Style="{StaticResource Text-H2}"
+                                   Text="{Binding Rank}" />
+                    </Border>
+                    <!-- Athlete / Horse -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <StackPanel Style="{StaticResource List-Vertical}">
+                            <!-- Athlete -->
+                            <TextBlock Style="{StaticResource Text-Bold}">
+                                <Run Text="{Binding ParticipantNumber}" />
+                                <Run Text="/" />
+                                <Run Text="{Binding AthleteName}" />
+                                <Run Style="{StaticResource Run}"
+                                     Text="{Binding AthleteFeiId}" />
+                            </TextBlock>
+                            <!-- Horse -->
+                            <TextBlock Style="{StaticResource Text}">
+                                <Run Text="{Binding HorseName}" />
+                                <Run Text=" / " />
+                                <Run Text="{Binding HorseFeiId}" />
+                                <Run Text=" / " />
+                                <Run Text="{Binding HorseGenderString}" />
+                            </TextBlock>
+                            <!-- Trainer -->
+                            <TextBlock Style="{StaticResource Text-Small}">
+                                <Run Text="Trainer : " />
+                                <Run Text="{Binding TrainerFeiId}" />
+                                <Run Text=" " />
+                                <Run Text="{Binding TrainerName}" />
+                            </TextBlock>
+                        </StackPanel>
+                    </Border>
+                    <!-- Country -->
+                    <Border Style="{StaticResource Border-Horizontal}">
+                        <TextBlock Style="{StaticResource Text-H2}"
+                                   Text="{Binding AthleteCountryName}" />
+                    </Border>
+                </StackPanel>
+
+                <!-- // TODO:
+                1. Create total column and row
+                2. Add static (ParticipantTemplate) labels column
+                3. Add list of Performances (Performance templates)
+                 -->
+
+                <!-- Columns (Headers) -->
                 <StackPanel Style="{StaticResource List-Horizontal}">
 
+                    <!-- Number -->
                     <Border Style="{StaticResource Border-Horizontal}" Width="150px">
                         <TextBlock Style="{StaticResource Text-H2}">
                             <Run>#</Run>
@@ -15,6 +64,7 @@
                         </TextBlock>
                     </Border>
 
+                    <!-- Phases -->
                     <Border Style="{StaticResource Border-Horizontal}">
                         <ListBox Style="{StaticResource ListBox-Horizontal}" BorderThickness="0"
                                  ItemsSource="{Binding PhaseLengths}">
@@ -31,6 +81,7 @@
                         </ListBox>
                     </Border>
 
+                    <!-- Collapse/Expand Button -->
                     <Border Style="{StaticResource Border-Horizontal}">
                         <Button Style="{StaticResource Button-Horizontal}"
                                 Command="{Binding ToggleVisibility}"
@@ -38,8 +89,10 @@
                     </Border>
                 </StackPanel>
 
+                <!-- Performances -->
                 <StackPanel Style="{StaticResource List-Horizontal}" Visibility="{Binding Visibility}">
 
+                    <!-- Labels -->
                     <Border Style="{StaticResource Border-Horizontal}" Width="150px">
                         <StackPanel Style="{StaticResource List-Vertical}">
 
@@ -95,12 +148,14 @@
                         </StackPanel>
                     </Border>
 
+                    <!-- Performances -->
                     <Border Style="{StaticResource Border-Horizontal}">
                         <ListBox Style="{StaticResource ListBox-Horizontal}" BorderThickness="0"
                                  ItemsSource="{Binding Performances}"
                                 Name="Performances" />
                     </Border>
 
+                    <!-- Print button -->
                     <Border Style="{StaticResource Border-Horizontal}">
                         <Button Style="{StaticResource Button-Big-Horizontal}"
                                 Command="{Binding Print}"
